Issue #ED-3963 fix: Updated build config and removed unwanted files
