--- MainActivity.java
@@ -1,7 +1,0 @@
-package org.sunbird.app;
-/**
- * Created by swayangjit on 16/10/19.
- */
-import com.getcapacitor.BridgeActivity;
-
-public class MainActivity extends BridgeActivity {}
