+++ android/app/src/main/res/values/splash_styles.xml
