+++ android/app/src/main/res/values/whitelabel.xml
