--- build_config
@@ -1,46 +1,0 @@
-cordova-plugin=cordova-plugin-printer
-cordova-plugin=https://github.com/Sunbird-Ed/jjdltc-cordova-plugin-zip.git
-cordova-plugin=ionic-plugin-keyboard
-cordova-plugin=cordova-plugin-ionic-webview
-cordova-plugin=cordova-plugin-app-version
-cordova-plugin=cordova-plugin-x-socialsharing
-cordova-plugin=cordova-plugin-androidx-adapter
-cordova-plugin=cordova-plugin-device
-cordova-plugin=cordova-plugin-file
-cordova-plugin=https://github.com/shikshalokam/cordova-plugin-file-transfer.git
-cordova-plugin=cordova-plugin-inappbrowser
-cordova-plugin=cordova-plugin-network-information
-cordova-plugin=cordova-plugin-statusbar
-cordova-plugin=cordova-plugin-webview-checker
-cordova-plugin=https://github.com/adriano-di-giovanni/cordova-plugin-shared-preferences.git
-cordova-plugin=https://github.com/swayangjit/cordova-plugin-local-notifications.git
-cordova-plugin=https://github.com/Sunbird-Ed/sb-cordova-plugin-fcm.git#release-5.1.2
-cordova-plugin=cordova-plugin-advanced-http
-cordova-plugin=cordova-plugin-android-permissions
-cordova-plugin=cordova.plugins.diagnostic
-cordova-plugin=com.telerik.plugins.nativepagetransitions
-cordova-plugin=cordova-plugin-secure-storage-echo
-cordova-plugin=https://github.com/shikshalokam/cordova-plugin-chooser.git
-cordova-plugin=cordova-plugin-appavailability
-cordova-plugin=cordova-plugin-filepicker
-cordova-plugin=cordova-plugin-document-viewer
-cordova-plugin=com-sarriaroman-photoviewer
-cordova-plugin=https://github.com/shikshalokam/cordova-plugin-filepath.git
-cordova-plugin=https://github.com/greybax/cordova-plugin-proguard.git
-cordova-plugin=https://github.com/Sunbird-Ed/sb-cordova-plugin-utility/tree/release-6.0.0
-cordova-plugin=https://github.com/Sunbird-Ed/sb-cordova-plugin-downloadmanager.git
-cordova-plugin=https://github.com/Sunbird-Ed/sb-cordova-plugin-db.git#release-4.6.0
-cordova-plugin=https://github.com/Sunbird-Ed/sb-cordova-plugin-customtabs#release-5.0.2 --variable URL_SCHEME=@string/custom_scheme_url --variable URL_HOST=mobile
-cordova-plugin=cordova-sqlite-storage
-cordova plugin=cordova-plugin-googleplus
-cordova-plugin=https://github.com/shikshalokam/cordova-plugin-camera.git
-cordova-plugin=cordova-plugin-telerik-imagepicker@2.3.3
-cordova-plugin=cordova-plugin-filechooser
-cordova-plugin=cordova-plugin-media
-cordova-plugin=cordova.plugins.diagnostic
-sunbird-cordova-plugin=https://github.com/project-sunbird/cordova-plugin-openrap.git
-sunbird-cordova-plugin=https://github.com/project-sunbird/cordova-plugin-qr-scanner.git#release-5.0.2
-sunbird-cordova-plugin=https://github.com/project-sunbird/cordova-plugin-sunbirdsplash.git#release-6.0.0
-sunbird-cordova-plugin=https://github.com/project-sunbird/sb-cordova-plugin-sync.git
-sunbird-cordova-plugin=https://github.com/subranil/cordova-plugin-inappupdatemanager.git#release-3.7.0
-cordova-plugin=uk.co.workingedge.phonegap.plugin.istablet
