--- configurations/android/colors.xml
@@ -1,4 +1,0 @@
-<?xml version='1.0' encoding='utf-8'?>
-<resources xmlns:tools="http://schemas.android.com/tools">
-    <color name="cdv_splashscreen_background">#ffffff</color>
-</resources>
