--- configurations/android/themes.xml
@@ -1,9 +1,0 @@
-<?xml version='1.0' encoding='utf-8'?>
-<resources>
-    <style name="Theme.App.SplashScreen" parent="Theme.SplashScreen.Common">
-        <item name="windowSplashScreenBackground">@color/cdv_splashscreen_background</item>
-        <item name="windowSplashScreenAnimatedIcon">@mipmap/ic_launcher</item>
-        <item name="windowSplashScreenAnimationDuration">5000</item>
-        <item name="postSplashScreenTheme">@style/Theme.AppCompat.NoActionBar</item>
-    </style>
-</resources>
