+++ package.json
@@ -13,8 +13,8 @@
     "test:coverage": "jest --coverage",
     "test:ci": "jest --ci --collectCoverage=true --coverageReporters=lcov --forceExit --detectOpenHandles --maxWorkers=2",
     "updateLink": "tsc && npm rm node_modules/@project-sunbird/sunbird-sdk/ -g && npm link ../../CapacitorProj/sunbird-mobile-sdk_cap/tmp/",
-    "capacitor:sync:after": "node hooks/capacitor-copy-after.js android && node scripts/application/add_application.js android && node hooks/capacitor-update-manifest.js android",
-    "capacitor:copy:before": "node scripts/copyContentPlayer.js android && node hooks/copyGradleProperties.js android && node scripts/android/fix-android-sdk33-migration.js android",
+    "capacitor:sync:after": "node scripts/copyGoogleServiceFile.js android && node scripts/application/add_application.js android",
+    "capacitor:copy:before": "node scripts/copyContentPlayer.js android && node scripts/copyGradleProperties.js android && node scripts/android/fix-android-sdk33-migration.js android",
     "capacitor:copy:after": "node scripts/copyContentPlayer.js android && node scripts/android/fix-android-sdk33-migration.js android",
     "ionic-build": "ng build && ionic cap run android --source-map=true --target Pixel_3_API_31",
     "ionic-build:prod": "ng build --configuration=$npm_config_angular_configuration && cordova build android --prod --release --verbose --buildConfig ./buildConfig/build.json",
@@ -29,6 +29,7 @@
     "@angular/animations": "15.2.10",
     "@angular/cdk": "15.2.9",
     "@angular/common": "15.2.10",
+    "@angular/compiler": "15.2.10",
     "@angular/core": "15.2.10",
     "@angular/forms": "15.2.10",
     "@angular/localize": "15.2.10",
@@ -70,8 +71,8 @@
     "@capawesome/capacitor-file-picker": "5.1.1",
     "@ionic/angular": "6.7.5",
     "@ionic/storage": "4.0.0",
-    "@ngx-translate/core": "11.0.1",
-    "@ngx-translate/http-loader": "4.0.0",
+    "@ngx-translate/core": "14.0.0",
+    "@ngx-translate/http-loader": "6.0.0",
     "@project-sunbird/client-services": "7.0.6",
     "@project-sunbird/common-consumption": "8.0.1",
     "@project-sunbird/common-form-elements": "8.0.5",
@@ -81,7 +82,7 @@
     "@project-sunbird/sb-notification": "8.0.0",
     "@project-sunbird/sb-styles": "0.0.16",
     "@project-sunbird/sb-svg2pdf": "8.0.4",
-    "@project-sunbird/sunbird-epub-player-web-component": "^1.2.0",
+    "@project-sunbird/sunbird-epub-player-web-component": "1.2.0",
     "@project-sunbird/sunbird-pdf-player-web-component": "1.1.0",
     "@project-sunbird/sunbird-quml-player-web-component": "3.0.3",
     "@project-sunbird/sunbird-sdk": "7.0.21",
@@ -98,7 +99,7 @@
     "datatables.net-fixedcolumns": "3.3.3",
     "dayjs": "1.9.8",
     "dom-to-image": "2.6.0",
-    "epubjs": "0.3.89",
+    "epubjs": "0.3.93",
     "grapheme-splitter": "1.0.4",
     "hammerjs": "2.0.8",
     "ionic4-rating": "1.0.9",
@@ -151,7 +152,6 @@
   "devDependencies": {
     "@angular-devkit/build-angular": "15.2.10",
     "@angular/cli": "15.2.10",
-    "@angular/compiler": "15.2.10",
     "@angular/compiler-cli": "15.2.10",
     "@angular/language-service": "15.2.10",
     "@angularclass/hmr": "3.0.0",
@@ -166,8 +166,6 @@
     "babel-polyfill": "6.26.0",
     "code-push": "3.0.1",
     "com-sarriaroman-photoviewer": "1.3.0",
-    "com.telerik.plugins.nativepagetransitions": "0.7.0",
-    "cordova-android": "12.0.0",
     "cordova-plugin-advanced-http": "2.5.1",
     "cordova-plugin-android-downloadmanager": "git+https://github.com/Sunbird-Ed/sb-cordova-plugin-downloadmanager.git",
     "cordova-plugin-code-push": "git+https://github.com/swayangjit/cordova-plugin-code-push.git",
