+++ scripts/copyContentPlayer.js
@@ -1,4 +1,3 @@
-const path = require('path');
 const fs = require('fs-extra');
 
 function copyFromNodeModule(src, dist) {
@@ -26,25 +25,23 @@ function copyFromContentPlayer(src, dist) {
     }
 
 }
-// module.exports = function (context) {
-    var srcPath;
-    var destinationPath;
-    var destinationPath1;
-    console.log('***** copy content');
-    srcPath = path.join(__dirname, '../content-player');
-    destinationPath = path.join(__dirname, '../www/content-player');
-    destinationPath1 = path.join(__dirname, '../dist/content-player');
-    copyFromContentPlayer(srcPath, destinationPath);
-    copyFromContentPlayer(srcPath, destinationPath1);
-    console.log('copied from content-player to www/content-player');
+var srcPath;
+var destinationPath;
+var destinationPath1;
+console.log('***** copy content');
+srcPath = '../content-player';
+destinationPath = '../www/content-player';
+destinationPath1 = '../dist/content-player';
+copyFromContentPlayer(srcPath, destinationPath);
+copyFromContentPlayer(srcPath, destinationPath1);
+console.log('copied from content-player to www/content-player');
 
-    srcPath = path.join(__dirname, '../node_modules/@project-sunbird/content-player');
-    destinationPath = path.join(__dirname, '../www/content-player');
-    destinationPath1 = path.join(__dirname, '../dist/content-player');
-    copyFromNodeModule(srcPath, destinationPath);
-    copyFromNodeModule(srcPath, destinationPath1);
-    console.log('copied from node_modules/content-player to wwww/content-player');
-// }
+srcPath = '../node_modules/@project-sunbird/content-player';
+destinationPath = '../www/content-player';
+destinationPath1 = '../dist/content-player';
+copyFromNodeModule(srcPath, destinationPath);
+copyFromNodeModule(srcPath, destinationPath1);
+console.log('copied from node_modules/content-player to wwww/content-player');
