+++ scripts/uploadAppIcon.js
